Update patches (#40)
This commit makes the following changes:
- `cna_order_queue` is not manually unrolled twice, but rather we use
`__VERIFIER_loop_bound`. Since the loop is bounded by the number of
threads, this means that we are not removing any code.
- It reverts the change to use marked accesses to initialize the node.
This change was introduced when we were still using GenMC.
- It removes `FIX1` because,
[as stated by Alan Stern](https://lkml.org/lkml/2022/9/12/447), 
`FIX2` should also solve the liveness bug.

The verification times are expected to be larger now.